requirements.txt
amqp==5.3.1
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==25.11.0
bleach==6.2.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
django==4.2.29
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.24.2
fqdn==1.5.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.2
identify==2.6.15
idna==3.11
importlib-metadata==8.7.1
ipykernel==6.31.0
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kombu==5.6.2
lark==1.3.1
lxml==6.0.2
markupsafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
mypy-extensions==1.1.0
nbclient==0.10.2
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.10.0
notebook==7.5.5
notebook-shim==0.2.4
numpy==2.0.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==11.3.0
pip==26.0.1
platformdirs==4.4.0
pre-commit==4.3.0
prometheus-client==0.24.1
prompt-toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pygments==2.19.2
pyjwt==2.11.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-json-logger==4.0.0
pytokens==0.4.1
pytz==2026.1.post1
pyyaml==6.0.3
pyzmq==27.1.0
redis==7.3.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
ruff==0.15.4
safetensors==0.7.0
send2trash==2.1.0
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
sqlalchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
vine==5.1.0
virtualenv==21.1.0
wcwidth==0.6.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
zipp==3.23.0
